Move thread into activity to stop it when the activity is destroy
