+++ example-databinding/src/main/java/com/xwray/groupie/example/databinding/item/LiveDataItem.java
@@ -6,9 +6,6 @@ import com.xwray.groupie.databinding.BindableItem;
 import com.xwray.groupie.example.databinding.R;
 import com.xwray.groupie.example.databinding.databinding.ItemLiveDataBinding;
 
-import java.util.Random;
-
-
 public class LiveDataItem extends BindableItem<ItemLiveDataBinding> {
 
     // Simple wrapper around liveData
@@ -21,25 +18,10 @@ public class LiveDataItem extends BindableItem<ItemLiveDataBinding> {
         }
     }
 
-    private DataWrapper internalDataWrapper = new DataWrapper();
+    public DataWrapper internalDataWrapper = new DataWrapper();
 
     public LiveDataItem() {
         super();
-        // simple thread which postValue in the liveData every 500ms
-        new Thread(new Runnable() {
-            @Override
-            public void run() {
-                Random random = new Random();
-                while(true) {
-                    internalDataWrapper.data.postValue(String.valueOf(random.nextLong()));
-                    try {
-                        Thread.sleep(500);
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-        }).start();
     }
 
     @Override public int getLayout() {
@@ -48,6 +30,5 @@ public class LiveDataItem extends BindableItem<ItemLiveDataBinding> {
 
     @Override public void bind(@NonNull ItemLiveDataBinding viewBinding, int position) {
         viewBinding.setWrapper(internalDataWrapper);
-        viewBinding.executePendingBindings();
     }
 }
